+++ src/lib/services/searchService.ts
@@ -7,7 +7,7 @@ import {
   useTierConfig,
   useValidateAddress,
 } from "lib/app-provider";
-import type { BechAddr, Option } from "lib/types";
+import type { Addr, BechAddr, Nullish, Option } from "lib/types";
 import { zBechAddr32, zValidatorAddr } from "lib/types";
 import type { IcnsNamesByAddress } from "lib/types/name";
 import {
@@ -18,12 +18,16 @@ import {
   splitModule,
 } from "lib/utils";
 
-import { useBlockData } from "./block";
+import { useBlockData, useBlockDataLcd } from "./block";
 import { useModuleByAddressLcd } from "./move/module";
 import { useAddressByIcnsNameLcd, useIcnsNamesByAddressLcd } from "./name";
 import { usePoolByPoolId } from "./poolService";
 import { useProposalData, useProposalDataLcd } from "./proposal";
 import { useTxData } from "./tx";
+import {
+  useAddressByInitiaUsername,
+  useInitiaUsernameByAddress,
+} from "./username";
 import { useValidatorData, useValidatorDataLcd } from "./validator";
 import { useCodeLcd } from "./wasm/code";
 import { useContractData } from "./wasm/contract";
@@ -45,18 +49,25 @@ export interface ResultMetadata {
     address: Option<BechAddr>;
     bech32Prefix: string;
   };
+  initiaUsername: {
+    username: Nullish<string>;
+    address: Nullish<Addr>;
+    bech32Prefix: string;
+  };
+}
+
+interface SearchHandlerResponse {
+  results: SearchResultType[];
+  isLoading: boolean;
+  metadata: ResultMetadata;
 }
 
 // eslint-disable-next-line complexity
 export const useSearchHandler = (
   keyword: string,
   resetHandlerStates: () => void
-): {
-  results: SearchResultType[];
-  isLoading: boolean;
-  metadata: ResultMetadata;
   // eslint-disable-next-line sonarjs/cognitive-complexity
-} => {
+): SearchHandlerResponse => {
   const isFullTier = useTierConfig() === "full";
   const [debouncedKeyword, setDebouncedKeyword] = useState(keyword);
   const {
@@ -107,18 +118,36 @@ export const useSearchHandler = (
   ) as BechAddr;
   const { data: icnsNames } = useIcnsNamesByAddressLcd(address);
 
+  // Initia Username
+  const {
+    data: initiaUsernameAddrData,
+    isFetching: initiaUsernameAddrFetching,
+  } = useAddressByInitiaUsername(debouncedKeyword);
+
+  const { data: initiaUsername, isFetching: initiaUsernameFetching } =
+    useInitiaUsernameByAddress(
+      isAddr ? (debouncedKeyword as Addr) : initiaUsernameAddrData?.address,
+      isMove
+    );
+
   const addressResult = useMemo(() => {
     if (isAddr) {
+      // eslint-disable-next-line sonarjs/no-duplicate-string
       return contractData ? "Contract Address" : "Account Address";
     }
-    return (
-      icnsAddressData?.address &&
-      (icnsAddressData.addressType === "contract_address"
-        ? "Contract Address"
-        : "Account Address")
-    );
-  }, [isAddr, contractData, icnsAddressData]);
 
+    if (icnsAddressData?.address)
+      return (
+        icnsAddressData.address &&
+        (icnsAddressData.addressType === "contract_address"
+          ? "Contract Address"
+          : "Account Address")
+      );
+
+    if (initiaUsernameAddrData?.address) return "Account Address";
+
+    return false;
+  }, [isAddr, icnsAddressData, initiaUsernameAddrData, contractData]);
   // Code
   const { data: codeData, isFetching: codeFetching } = useCodeLcd(
     Number(debouncedKeyword),
@@ -135,7 +164,7 @@ export const useSearchHandler = (
     Number(debouncedKeyword),
     isPosDecimal(debouncedKeyword) && isFullTier
   );
-  const blockLcd = useBlockData(
+  const blockLcd = useBlockDataLcd(
     Number(debouncedKeyword),
     isPosDecimal(debouncedKeyword) && !isFullTier
   );
@@ -237,6 +266,8 @@ export const useSearchHandler = (
       contractFetching ||
       blockFetching ||
       icnsAddressFetching ||
+      initiaUsernameAddrFetching ||
+      initiaUsernameFetching ||
       poolFetching ||
       proposalFetching ||
       validatorFetching,
@@ -244,6 +275,13 @@ export const useSearchHandler = (
       icns: {
         icnsNames,
         address,
+        bech32Prefix,
+      },
+      initiaUsername: {
+        username: initiaUsername?.username,
+        address: isAddr
+          ? (debouncedKeyword as Addr)
+          : initiaUsernameAddrData?.address,
         bech32Prefix,
       },
     },